backend/requirements.txt
# FastAPI 后端依赖包
# 使用命令安装: pip install -r requirements.txt

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# 配置管理
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 图片处理
Pillow==10.2.0
piexif==1.1.3

# AI调用
httpx==0.26.0
openai==1.10.0

# 工具库
python-dateutil==2.8.2
